Test/requirements.txt
selenium
webdriver-manager
requests
python-dotenv
google-generativeai
